解决所以警告

+++ AV-CSG/EnemyBullet.cpp
@@ -82,8 +82,8 @@ void CEnemyBullet::Update()
     float fDis = tD * m_nSpeed;
     int nStartX = 0, nStartY = 0;
 
-    m_nPosX += cos(m_fAngle) * fDis;
-    m_nPosY += sin(m_fAngle) * fDis;
+    m_nPosX += int(cos(m_fAngle) * fDis);
+    m_nPosY += int(sin(m_fAngle) * fDis);
 
     m_nFrameStartX = m_nCurrentFrame * m_nWidth;
     ++m_nCurrentFrame;
